backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
resend==2.26.0
pypdf==4.3.1
reportlab==4.2.5
python-multipart==0.0.12
